% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_to_list.R
\name{append_to_list}
\alias{append_to_list}
\title{Append an item to a list dynamically (pipe-optimised)}
\usage{
append_to_list(x, list_x, name = "")
}
\arguments{
\item{x}{An object to append to list, e.g. vector, data frame.}

\item{list_x}{Target list to append object to.}

\item{name}{Specify a character string for the name of the list. Defaults to blank}
}
\description{
Function directly appends item to list in Global Environment
}
\examples{
a_list <- list(NULL)
append_to_list(iris,a_list,"iris") 
}
