% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_row.R
\name{apply_row}
\alias{apply_row}
\title{Apply a function rowwise, selecting variables with dplyr::select() syntax}
\usage{
apply_row(x, select_helpers = everything(), FUN, ...)
}
\arguments{
\item{x}{Data frame or tibble to pass through.}

\item{select_helpers}{Select variables using dplyr::select() syntax}

\item{FUN}{Function to be applied to selected columns}
}
\description{
Apply a function rowwise, selecting variables with dplyr::select() syntax
}
\examples{
\dontrun{
library(tidyverse)
iris \%>\% mutate(Sepal_Sum = apply_row(.,select_helpers = contains("Sepal"), sum, na.rm = TRUE))
iris \%>\% mutate(Sepal_Sum = apply_row(.,c("Sepal.Length","Petal.Width"), function(x) sum(x > 1))) 
}

}
