% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_nps.R
\name{as_nps}
\alias{as_nps}
\title{Convert numeric variable to NPS variable}
\usage{
as_nps(x)
}
\arguments{
\item{x}{Numeric variable to pass through. Valid range is 0 to 10 inclusive, otherwise returns a NA.}
}
\description{
Convert numeric variable to NPS variable
}
