% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_percent.R
\name{as_percent}
\alias{as_percent}
\title{Convert as percent (string)}
\usage{
as_percent(num, rounding = 0)
}
\arguments{
\item{num}{Numeric vector to pass through}

\item{rounding}{Number of decimal places to round to. Default is 0.}
}
\description{
Convert a numeric value into a string with percentage sign.
}
\examples{
as_percent(.86748)

}
