% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pc_loglin.R
\name{calc_pc_loglin}
\alias{calc_pc_loglin}
\title{Calculate percentage impact from coefficients of a log-linear model}
\usage{
calc_pc_loglin(x)
}
\arguments{
\item{x}{A log-linear model object.}
}
\value{
A \link[tibble:tibble-package]{tibble} with three columns: \code{var}, \code{coef}, and \code{pc_impact}.
}
\description{
This function exponentiates coefficients and takes out 1 to calculate the
percentage impact of each variable on the response variable in a log-linear
model. The function returns a tibble with three columns: \code{var}, \code{coef}, and
\code{pc_impact}.
}
