% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_weights.R
\name{calc_weights}
\alias{calc_weights}
\title{Calculate (simple) weights based on the proportion of another variable.}
\usage{
calc_weights(data, weight_var, target, weight_label = "weight")
}
\arguments{
\item{data}{Data frame or tibble containing the variable used for computing weights}

\item{weight_var}{The variable to be used for computing the weight}

\item{target}{A named vector containing the target weights. Names must mtach labels for `weight_var`}

\item{weight_label}{A character vector to specify the variable name of the weight variable. Defaults to 'weight'.}
}
\description{
Calculate (simple) weights based on the proportion of another variable.
}
