% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorise.R
\name{categorise}
\alias{categorise}
\title{Convert numeric variable into categorical variable}
\usage{
categorise(x, breaks)
}
\arguments{
\item{x}{A numeric vector to be passed through.}

\item{breaks}{Break points for the categories. Both the top and bottom breaks must be provided.}
}
\value{
Returns a character vector.
}
\description{
This is a wrapper around the cut() function.
}
\examples{
categorise(seq(10),breaks=c(0,3,10))
}
