% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_to_lab.R
\name{char_to_lab}
\alias{char_to_lab}
\title{Convert character variable to labelled integer variable}
\usage{
char_to_lab(x)
}
\arguments{
\item{x}{Character vector to pass through}
}
\description{
This function converts the character values into value labels, assigning each value an integer.
}
