% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_freq_dist.R
\name{create_freq_dist}
\alias{create_freq_dist}
\title{Create frequency distribution table for a metric}
\usage{
create_freq_dist(data, metric)
}
\arguments{
\item{data}{A data frame containing the data}

\item{metric}{The name of the metric for which the frequency distribution is
to be created}
}
\description{
This function creates a frequency distribution table for a given
metric. The table contains the bin ranges and the counts of the data points
that fall within each bin.
}
