% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fa_loads.R
\name{extract_fa_loads}
\alias{extract_fa_loads}
\title{Function to create a loadings file from the \code{stats::factanal()} output}
\usage{
extract_fa_loads(fa_object)
}
\arguments{
\item{fa_object}{factanal() model}
}
\description{
Function to create a loadings file from the \code{stats::factanal()} output
}
\examples{
\dontrun{
fa_output <- stats::factanal(
  tidyr::drop_na(psych::bfi),
  factors = 6
  )
extract_fa_loads(fa_output)
}

}
\keyword{analysis}
\keyword{factor}
