% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_vallab.R
\name{extract_vallab}
\alias{extract_vallab}
\title{Return value labels as tibble}
\usage{
extract_vallab(x, variable)
}
\arguments{
\item{x}{Data frame or tibble containing the variable}

\item{variable}{Enter variable name as string}
}
\description{
This function is conditional on the variable having value labels.
}
