% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factanal functions.r
\name{fa.score}
\alias{fa.score}
\title{Function to create a score file from the factanal() output}
\usage{
fa.score(x, file_name = NULL, writeCSV = FALSE)
}
\arguments{
\item{x}{factanal() model}
}
\description{
Function to create a score file from the factanal() output
}
\keyword{analysis}
\keyword{factor}
