% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclustfunc.R
\name{hclustfunc}
\alias{hclustfunc}
\title{Run hierarchical clustering with arguments to specify methods.}
\usage{
hclustfunc(x, method = "complete", dmeth = "euclidean")
}
\description{
Run hierarchical clustering with arguments to specify methods.
}
