% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_to_char.R
\name{lab_to_char}
\alias{lab_to_char}
\title{Convert labelled double variable to character variable}
\usage{
lab_to_char(x)
}
\arguments{
\item{x}{Labelled double vector to pass through}
}
\description{
Returns error if variable is not of class labelled double.
}
