% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_quantile.R
\name{labelled_quantile}
\alias{labelled_quantile}
\title{Convert a numeric into a quantile categorical variable, labelled by lower and upper bounds of quantiles (string)}
\usage{
labelled_quantile(x)
}
\arguments{
\item{x}{Vector, column, or numeric variable to be passed through}
}
\description{
Convert a numeric into a quantile categorical variable, labelled by lower and upper bounds of quantiles (string)
}
\examples{
labelled_quantile(c(1,1,1,1,2,3,5,5,6))
}
