% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_convert.R
\name{likert_convert}
\alias{likert_convert}
\title{Convert a Likert scale from one scale to another}
\usage{
likert_convert(x, top.x, bot.x, top.y, bot.y)
}
\arguments{
\item{x}{Numeric vector to be passed through.}

\item{top.x}{Top value of the original scale. This would be 6 on a 0-6 scale.}

\item{bot.x}{Bottom value of the original scale. This would be 0 on a 0-6 scale.}

\item{top.y}{Top value of the new/target scale. This would be 5 on a 0-5 scale.}

\item{bot.y}{Bottom value of the new/target scale. This would be 0 on a 0-5 scale.}
}
\value{
A numeric vector containing the rescaled likert scale variable.
}
\description{
This is used for converting data, for instance, from a 6-point scale to a 5-point scale.
The scale "dimensions" are specified in the function argument, and the function converts the numeric vector that is passed through.
}
\examples{
data <-c(5, 4, 3, 2, 1)
likert_convert(data,5,0,10,0) #5-point scale to 10-point scale
[1] 10  8  6  4  2
}
\keyword{likert}
\keyword{scale}
