% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_reverse.R
\name{likert_reverse}
\alias{likert_reverse}
\title{Reverse a Likert scale}
\usage{
likert_reverse(x, top, bottom)
}
\arguments{
\item{x}{Numeric vector}

\item{top}{Top value of the scale for the variable, e.g. 10 for a 0-10 scale.}

\item{bottom}{value of the scale for the variable, e.g. 0 for a 0-10 scale.}
}
\value{
A numeric vector which is the "reverse" of the original likert scale variable.
}
\description{
Reverse a Likert scale such that on a 0-10 scale, 10 becomes 0 and 0 becomes 10. 
Use the `top` and `bottom` arguments to specify the valid range of the scale.
}
\examples{
data <-c(5, 4, 3, 2, 6)
likert_reverse(data, 6, 0)
}
\keyword{likert}
\keyword{scale}
