% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxmin.R
\name{maxmin}
\alias{maxmin}
\title{Max-Min Scaling Function}
\usage{
maxmin(x)
}
\arguments{
\item{x}{Pass a vector or the required columns of a data frame through this argument.}
}
\description{
This function allows you to scale vectors or an entire data frame using the max-min scaling method
A numeric vector is always returned.
}
\examples{
library(magrittr)
library(dplyr)


rand.data <- data.frame(x = sample(x = 1000, size = 234:697))
rand.data \%>\%
  mutate(
    y = maxmin(x)
  )

sample(x = 1000, size = 234:677) \%>\% maxmin()

iris \%>\% mutate(Petal.Length2 = maxmin(Petal.Length))

maxmin(iris$Petal.Length)
}
\keyword{max-min}
