% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_change.R
\name{pc_change}
\alias{pc_change}
\title{Calculate percentage change of a vector relative to a lag k}
\usage{
pc_change(x, lag = 1)
}
\arguments{
\item{x}{A vector to be passed through}

\item{lag}{The number of lags used, defaults to 1}
}
\description{
Calculate percentage change of a vector relative to a lag k
}
