% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factanal functions.r
\name{plain_num}
\alias{plain_num}
\title{Function to convert all relevant columns to numeric
Create a plain data frame without GUID and category}
\usage{
plain_num(x)
}
\arguments{
\item{x}{data frame to pass through}
}
\description{
Function to convert all relevant columns to numeric
Create a plain data frame without GUID and category
}
\keyword{convert}
\keyword{numeric}
\keyword{to}
