% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_vallab.R
\name{recode_vallab}
\alias{recode_vallab}
\title{Recode value labels based on numeric code}
\usage{
recode_vallab(x, code, new_label)
}
\arguments{
\item{x}{haven_labelled vector to be passed through.}

\item{code}{Numeric value that specifies code where value label is to be replaced.}

\item{new_label}{Character string to be used as the new value label.}
}
\description{
Recode value labels based on numeric code
}
\examples{
x <- haven::labelled(c(1,2,3,2,1),c("a" = 1, "b" = 2, "d" = 3))
recode_vallab(x, code = 2, new_label = "Surprise!")

}
