% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_na_only.R
\name{remove_na_only}
\alias{remove_na_only}
\title{Remove all columns which contain only NAs
Optimised for magrittr / dplyr pipes}
\usage{
remove_na_only(df)
}
\arguments{
\item{df}{Data frame to be passed through}
}
\description{
Remove all columns which contain only NAs
Optimised for magrittr / dplyr pipes
}
