% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_zero_only.R
\name{remove_zero_only}
\alias{remove_zero_only}
\title{Remove all columns which contain only zeros
Optimised for magrittr / dplyr pipes}
\usage{
remove_zero_only(df)
}
\arguments{
\item{df}{Data frame to be passed through}
}
\description{
Remove all columns which contain only zeros
Optimised for magrittr / dplyr pipes
}
