% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na_range.R
\name{replace_na_range}
\alias{replace_na_range}
\title{Replace NAs randomly from a selected range with replacement}
\usage{
replace_na_range(x, range)
}
\arguments{
\item{x}{Original numeric vector containing missing values.}

\item{range}{Vector specifying the valid values to replace the missing values with}
}
\description{
Replace NAs randomly from a selected range with replacement
}
\examples{
replace_na_range(c(1,NA,2,3,1,NA,2),c(1,2,3)) 

}
