% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hclust.R
\name{run_hclust}
\alias{run_hclust}
\title{A wrapper function to run hierarchical clustering}
\usage{
run_hclust(x, method = "complete", dmeth = "euclidean")
}
\arguments{
\item{x}{Data frame to run hierarchical clustering}

\item{method}{Specify method for hierarchical clustering. See ?hclust for details}

\item{dmethod}{Specify method for calculating distance. See ?dist for details}
}
\description{
A wrapper function to run hierarchical clustering
}
\examples{
library(magrittr)
library(dplyr)
iris \%>\% select(-Species) \%>\% run_hclust()
}
