% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sav_to_rds.R
\name{sav_to_rds}
\alias{sav_to_rds}
\title{Serialise a SPSS file to RDS}
\usage{
sav_to_rds(import, export = "")
}
\arguments{
\item{import}{String containing path to .SAV file}

\item{export}{String containing desired file name for RDS export.
Default is to use same name as .SAV file}
}
\value{
NULL, invisibly. The function is called for its side effect of
creating an RDS file.
}
\description{
This function is useful for converting large, bulky SAV files into RDS files which are
faster to load and take up less memory. Use readRDS() for loading in the exported file.
By default, the old file name is retained and saved in the same location as RDS
unless otherwise specified by the export argument.
}
