% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_vall.R
\name{set_vall}
\alias{set_vall}
\title{Set value labels}
\usage{
set_vall(x, value_labels)
}
\arguments{
\item{x}{Variable to assign value labels to}

\item{value_labels}{Named character vector to be assigned as value labels}
}
\value{
The input variable \code{x} as a labelled vector with value labels set.
}
\description{
A pipe-workflow optimised method to set value labels.
This is a wrapper around \code{haven::labelled()}, but preserves
the original variable label of the vector.
}
\examples{

library(magrittr)
library(dplyr)
library(tibble)
tibble(RESPID = 1:1000,
       Q1 = sample(c(0, 1, 2), 1000, replace = TRUE),
       Q2 = sample(c(0, 1), 1000, replace = TRUE)) -> df
df \%>\%
  mutate(Q2 = set_varl(Q2, "What is your answer to this yes/no question?")) \%>\%
  mutate(Q2 = set_vall(Q2, c("No" = 0, "Yes" = 1))) \%>\%
  .$Q2 \%>\% attributes()
}
\seealso{
Other Labels: 
\code{\link{set_varl}()}
}
\concept{Labels}
\concept{Value labels}
