% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_tt.R
\name{split_tt}
\alias{split_tt}
\title{Split the data into a simple training and testing set}
\usage{
split_tt(x, part)
}
\arguments{
\item{x}{Pass your data frame or matrix here.}

\item{part}{A numeric value between 0 and 1 to represent the proportion of the whole data you want to use as the training set.}
}
\description{
Split the data into a simple training and testing set
}
\examples{
x <-as.data.frame(matrix(1:5000,250,20))
x.train <- split_tt(x,.7)$train
x.test <- split_tt(x,.7)$test

dim(x.train)
dim(x.test)
}
\keyword{test}
\keyword{train,}
