% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_replace_nth.R
\name{str_replace_nth}
\alias{str_replace_nth}
\title{Replace nth occurring value in a string}
\usage{
str_replace_nth(x, pattern, replacement, n)
}
\arguments{
\item{x}{Pass string character}

\item{pattern}{String containing characters to match}

\item{replacement}{String containing to replace}

\item{n}{Nth term to be replaced}
}
\description{
A more refined replacement function than str_replace() and str_replace_all()
}
