% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_spread.R
\name{super_spread}
\alias{super_spread}
\title{Enhanced version of superspread() that creates dummy variables from multiple categorical variables.
Uses data.table() for speed.}
\usage{
super_spread(df, select_helpers)
}
\arguments{
\item{df}{Data frame containing the multiple categorical variables.}

\item{select_helpers}{Uses dplyr-style select functions to select multiple variables. 
Use everything() to select all variables.}
}
\description{
Enhanced version of superspread() that creates dummy variables from multiple categorical variables.
Uses data.table() for speed.
}
\examples{
library(data.table)
library(dplyr)
dt <- data.table(id = 1:10000,
cat1 = sample(letters[1:3],10000,replace = TRUE),
cat2 = sample(letters[1:4],10000,replace = TRUE),
cat3 = sample(letters[1:5],10000,replace = TRUE))
super_spread(dt, contains("cat"))

}
