% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superspread_count.R
\name{superspread_count}
\alias{superspread_count}
\title{Convert single-code column(s) into "multiple choice" formats, filling data with count rather than 1s}
\usage{
superspread_count(df, tcol, select_helpers = everything())
}
\arguments{
\item{df}{Data frame to apply function to.}

\item{tcol}{Number of columns to create}

\item{select_helpers}{Use dplyr select helpers to apply function only to required columns. 
Leave blank to include everything}
}
\description{
Convert single-code column(s) into "multiple choice" formats, filling data with count rather than 1s
}
\examples{
df <- data.frame(x=sample(1:50,1000,replace=TRUE))
superspread_count(df,50)

}
