% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_chisq.R
\name{test_chisq}
\alias{test_chisq}
\title{Compute chi-square or Fisher's exact test for two categorical variables}
\usage{
test_chisq(data, x, y, na_x = NULL, na_y = NULL)
}
\arguments{
\item{data}{A data frame containing the variables of interest.}

\item{x}{A character string specifying the name of the first variable.}

\item{y}{A character string specifying the name of the second variable.}

\item{na_x}{A vector of values to be treated as missing in \code{x}.}

\item{na_y}{A vector of values to be treated as missing in \code{y}.}
}
\value{
A tibble containing the results of the chi-square or Fisher's exact test.
}
\description{
This function computes a chi-square or Fisher's exact test for two categorical variables in a data frame.
}
\details{
If the cell counts are lower than 5, the function will use Fisher's exact test. Otherwise, it will use a chi-square test.
}
\examples{
data("mtcars")
test_chisq(mtcars, "cyl", "vs")

}
