% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timed_fn.R
\name{timed_fn}
\alias{timed_fn}
\title{Create file name with time stamp}
\usage{
timed_fn(main, extension)
}
\arguments{
\item{main}{The main file name to be used.}

\item{extension}{The file extension to be used, e.g. ".csv"}
}
\value{
A character string with the file name, timestamp, and extension combined.
}
\description{
This function generates a character string that suffixes a file name with a time stamp.
}
\examples{
timed_fn("Q15. ",".xlsx")
}
