% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenise_ngram.R
\name{tokenise_ngram}
\alias{tokenise_ngram}
\title{Return all n-gram combinations with a character vector as input}
\usage{
tokenise_ngram(x, n = 2, collapse = " ")
}
\arguments{
\item{x}{Vector of character string}

\item{n}{Specifying the n for n-grams, e.g. 2 for bi-grams}

\item{collapse}{Specify what to use for separating the outcome}
}
\description{
Return all n-gram combinations with a character vector as input
}
\examples{
tokenise_ngram("a c b e f")
[1] "a c" "c b" "b e" "e f"

}
