% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_nps.R
\name{ttest_nps}
\alias{ttest_nps}
\title{Performs a t-test on NPS}
\usage{
ttest_nps(x, conf_level = 0.95)
}
\arguments{
\item{x}{NPS vector to be passed through. Use \code{surveytoolbox::as_nps} to convert a 11-point scale into a NPS vector.}

\item{conf_level}{Numeric value to specify the confidence level. Defaults to 0.95.}
}
\description{
Returns a diagnostic message and the margin of error.
}
\examples{
x <- sample(0:10, 1000, replace = TRUE)
nps_x <- surveytoolbox::as_nps(x)
}
