% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varl_tb.R
\name{varl_tb}
\alias{varl_tb}
\title{Create tidy data frame with variable and variable labels}
\usage{
varl_tb(x)
}
\arguments{
\item{x}{Data frame containing labelled variables to pass through}
}
\description{
Input a data frame or tibble containing labelled variables,
and returns a tidy tibble with variable and variable labels.
Empty variables will be imputed with "No label"
}
