% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_x.R
\name{any_x}
\alias{any_x}
\title{Function that returns TRUE/FALSE if value exists in x, but returns NA if x consists entirely of NAs}
\usage{
any_x(x, value)
}
\description{
Function that returns TRUE/FALSE if value exists in x, but returns NA if x consists entirely of NAs
}
\examples{
any_x(c(1,0,1),1) # TRUE
any_x(c(1,NA,1),1) # TRUE
any_x(c(0,0,NA),1) # FALSE
any_x(c(NA,NA,NA),1) # NA

}
