% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_to_list.R
\name{append_to_list}
\alias{append_to_list}
\title{Append an item to a list dynamically}
\usage{
append_to_list(x, list_x, name = "", enviro = .GlobalEnv)
}
\arguments{
\item{x}{An object to append to list, e.g. vector, data frame.}

\item{list_x}{Target list to append object to.}

\item{name}{Specify a character string for the name of the list. Defaults to blank}

\item{enviro}{Specifies the environment}
}
\description{
The `append_to_list()` function appends an object to the specified list in Global Environment (default).
This function is pipe-optimised, and allows the option of specifying a name for the new object in the list.
}
\examples{
a_list <- list(NULL)
append_to_list(iris,a_list,"iris") 
}
