% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_it.R
\name{box_it}
\alias{box_it}
\title{Convert ordinal variables into binary variables by "boxing"}
\usage{
box_it(x, which = "top", number = 2)
}
\arguments{
\item{x}{Variable to be passed through}

\item{which}{Character string to specify which end of spectrum to take values - valid inputs are "top" and "bottom"}

\item{number}{Number to take values from}
}
\description{
For instance, you can create a Top Two Box variable from a 7-point agreement scale question.
Function returns a labelled double variable, which will have value label attributes.
}
\examples{
box_it(sample(1:10,100,replace = TRUE)) # Converted to binary variable where 9, 10 are selected
box_it(sample(c(1:10, NA),100,replace = TRUE))
}
