% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pc_loglin.R
\name{calc_pc_loglin}
\alias{calc_pc_loglin}
\title{Calculate percentage impact from coefficients of a log-linear model}
\usage{
calc_pc_loglin(x)
}
\arguments{
\item{x}{Log-linear model to be passed through}
}
\description{
Exponentiates coefficients and takes out 1 to calculate percentage impact.
Returns a tibble
}
