% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorise.R
\name{categorise}
\alias{categorise}
\title{Convert numeric variable into categorical variable}
\usage{
categorise(x, breaks)
}
\arguments{
\item{x}{vector to be passed through}

\item{break}{points for the categories. Both the top and bottom breaks must be provided.}
}
\description{
Essentially a wrapper around the cut() function
}
\examples{
seq(10) \%>\% categorise(breaks=c(0,3,10))
}
