% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_to_var.R
\name{chr_to_var}
\alias{chr_to_var}
\title{Convert a string / character variable into a labelled double variable,
using a pre-specified set of variable and value label mappings.}
\usage{
chr_to_var(var, lab_str, lab_num, var_label)
}
\arguments{
\item{var}{string vector containing the variable}

\item{lab_str}{string vector containing the labels for the variable. The
lengths and values must be an exact match of the variable in \code{var}.}

\item{lab_num}{numeric vector matching the labels provided in \code{lab_str}. The
length of \code{lab_num} must be the same as \code{lab_str}, and the order is used to
determine the mapping.}

\item{var_label}{string to be used as the variable label, passed through to
\code{set_varl()}.}
}
\value{
a labelled double variable
}
\description{
This function takes a string variable and converts it into a labelled double
variable, using value label mappings (\code{lab_str}, \code{lab_num}) and variable
labels (\code{var_label}) provided by the user.
}
\details{
This function is a wrapper around several other {surveytoolbox} functions:
\itemize{
\item \code{create_named_list()}
\item \code{set_vall()}
\item \code{set_varl()}
}
}
\examples{

var1 <- c("Yes", "No", "Not sure")
lab1 <- c(1, 0, 99)
q1_var <- sample(var1, 100, replace = TRUE)

chr_to_var(
  var = q1_var,
  lab_str = var1,
  lab_num = lab1,
  var_label = c("Have you come across this product previously?")
)

}
