% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_to_df.R
\name{cor_to_df}
\alias{cor_to_df}
\title{Convert correlation matrix into a tidy data frame
Use label_table as an input
Identify matching id}
\usage{
cor_to_df(cor_m, label_table, id)
}
\arguments{
\item{cor_m}{Correlation matrix in the form of an output from cor()}

\item{label_table}{A two-column table with one column used for matching and the other for returning labels.}

\item{id}{Use this argument to specify a string to refers to the matching / id column in the label_table.}
}
\description{
Convert correlation matrix into a tidy data frame
Use label_table as an input
Identify matching id
}
