% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_named_list.R
\name{create_named_list}
\alias{create_named_list}
\title{Create a named list object with two vectors}
\usage{
create_named_list(x, y)
}
\arguments{
\item{x}{Vector containing character strings, e.g. "A"}

\item{y}{Vector containing values for the list, e.g. 1, 3}
}
\description{
Input two vectors of the same length and return a named list.
}
\examples{
create_named_list(c("Alice","Bob","Carol"),c(54,60,23))
}
