% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dict.R
\name{data_dict}
\alias{data_dict}
\title{Create a Data Dictionary from a data frame with Variable and Value Labels
Exported objects from SPSS}
\usage{
data_dict(x, max_char = 32760)
}
\arguments{
\item{x}{data frame}

\item{max_char}{Maximum character allowed on Variable and Value label columns.}
}
\description{
Create a Data Dictionary from a data frame with Variable and Value Labels
Exported objects from SPSS
}
