% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look_up.R
\name{look_up}
\alias{look_up}
\title{Replace x values with corresponding values using a key}
\usage{
look_up(x, table, index = "var", column = 2)
}
\arguments{
\item{x}{A string vector to be matched.}

\item{table}{Lookup table}

\item{index}{Character string of the key/index column used for matching}

\item{column}{Column index (integer) or name (string) to return from the lookup table.
Default is the second column.}
}
\description{
Replace x values with corresponding values using a key
}
\examples{
library(magrittr)
library(dplyr)
iris \%>\% .$Petal.Width \%>\%
 look_up(tibble(var=0.2,return="zero-point-two"))
}
