% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxmin.R
\name{maxmin}
\alias{maxmin}
\title{Max-Min Scaling Function}
\usage{
maxmin(x)
}
\arguments{
\item{x}{Pass a vector or the required columns of a data frame through this argument.}
}
\description{
This function allows you to scale vectors or an entire data frame using the max-min scaling method
A numeric vector is always returned.
}
\examples{
rand.data <-cbind(sample(1000,234:697),sample(1000,234:697)) \%>\% as.data.frame()
rand.data \%>\% mutate_all(~maxmin(.))

sample(1000,234:677) \%>\% maxmin()

iris \%>\% mutate_at(vars(Petal.Length),~maxmin(.))
maxmin(iris$Petal.Length)
}
\keyword{max-min}
