% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxmin.R
\name{maxmin}
\alias{maxmin}
\title{Max-Min Scaling Function}
\usage{
maxmin(x)
}
\arguments{
\item{x}{Pass a vector or the required columns of a data frame through this argument.}
}
\description{
This function allows you to scale vectors or an entire data frame using the max-min scaling method,
always returning a data frame.
}
\examples{
rand.data <-cbind(sample(1000,234:697),sample(1000,234:697)) \%>\% as.data.frame()
maxmin(rand.data)

rand.data <-sample(1000,234:677)
maxmin(rand.data)
}
\keyword{max-min}
