% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_df.R
\name{read_df}
\alias{read_df}
\title{Read in a data frame in the clipboard, copied from Excel}
\usage{
read_df(header = TRUE, ...)
}
\arguments{
\item{header}{Logical value to specify whether copied table contains a header row.}

\item{...}{Additional arguments for read.table().}
}
\description{
Read in a data frame in the clipboard, copied from Excel
}
