% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sav_to_rds.R
\name{sav_to_rds}
\alias{sav_to_rds}
\title{Serialise a SPSS file to RDS
Old file name is retained and saved in the same location as RDS if export is unspecified}
\usage{
sav_to_rds(import, export = "")
}
\arguments{
\item{import}{String containing path to .SAV file}

\item{export}{String containing desired file name for RDS export. 
Default is to use same name as .SAV file}
}
\description{
Serialise a SPSS file to RDS
Old file name is retained and saved in the same location as RDS if export is unspecified
}
