% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superspread.R
\name{superspread}
\alias{superspread}
\title{Convert single-code column(s) into "multiple choice" formats - fill data with 1s}
\usage{
superspread(df, tcol, select_helpers = everything())
}
\arguments{
\item{df}{Data frame to apply function to.}

\item{tcol}{Number of columns to create}

\item{select_helpers}{Use dplyr select helpers to apply function only to required columns. 
Leave blank to include everything}
}
\description{
This function was orignally used in the context of a 'Select Any' question, where the outputs were created such that each 'cell' value represents the statement, rather than having the columns represent the statements. 
tcol refers to the number of columns to create. 
'Cell' values must be in numeric. 
Version with select incorporated
}
\examples{
df <- data.frame(x=sample(1:50,1000,replace=TRUE))
superspread(df,50)
}
