% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superspread.R
\name{superspread}
\alias{superspread}
\title{Creates dummy variables from multiple categorical variables.
Uses data.table() for speed (enhanced from the previous version)
Uses dplyr select() special functions to select categorical variables.}
\usage{
superspread(df, select_helpers)
}
\arguments{
\item{df}{Data frame containing the multiple categorical variables.}

\item{select_helpers}{Uses dplyr-style select functions to select multiple variables. 
Use everything() to select all variables. These variables must all be character type.}
}
\description{
Creates dummy variables from multiple categorical variables.
Uses data.table() for speed (enhanced from the previous version)
Uses dplyr select() special functions to select categorical variables.
}
\examples{
superspread(iris,"Species")
\dontrun{
library(data.table)
library(dplyr)
dt <- data.table(id = 1:10000,
                 cat1 = sample(letters[1:3],10000,replace = TRUE),
                 cat2 = sample(letters[1:4],10000,replace = TRUE),
                 cat3 = sample(letters[1:5],10000,replace = TRUE))
dt <- as.data.table(dt)
superspread(df = dt, select_helpers = contains("cat"))
}

}
\seealso{
Other superspread functions: \code{\link{superspread_count}}
}
\concept{superspread functions}
